refactoring Operation and Bulk statuses interfaces, remove illogical interfaces and methods
