+++ app/code/Magento/AsynchronousOperations/Api/Data/DetailedBulkStatusInterface.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+namespace Magento\AsynchronousOperations\Api\Data;
+
+/**
+ * Interface BulkStatusInterface
+ * Bulk summary data with list of operations items full data.
+ *
+ * @api
+ */
+interface DetailedBulkStatusInterface extends BulkSummaryInterface
+{
+
+    const OPERATIONS_LIST = 'operations_list';
+
+    /**
+     * Retrieve operations list.
+     *
+     * @return \Magento\AsynchronousOperations\Api\Data\OperationInterface[]
+     */
+    public function getOperationsList();
+
+    /**
+     * Set operations list.
+     *
+     * @param \Magento\AsynchronousOperations\Api\Data\OperationInterface[] $operationStatusList
+     * @return $this
+     */
+    public function setOperationsList($operationStatusList);
+}
