+++ app/code/Magento/AsynchronousOperations/etc/webapi.xml
@@ -8,15 +8,22 @@
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
 
-    <route url="/V1/bulk/status-detailed/:bulkUuid" method="GET">
-        <service class="Magento\AsynchronousOperations\Api\BulkRepositoryInterface" method="getBulkDetailedStatus"/>
+    <route url="/V1/bulk/:bulkUuid/detailed-status" method="GET">
+        <service class="Magento\AsynchronousOperations\Api\BulkStatusInterface" method="getBulkDetailedStatus"/>
         <resources>
             <resource ref="Magento_Logging::system_magento_logging_bulk_operations" />
         </resources>
     </route>
 
-    <route url="/V1/bulk/status-short/:bulkUuid" method="GET">
-        <service class="Magento\AsynchronousOperations\Api\BulkRepositoryInterface" method="getBulkShortStatus"/>
+    <route url="/V1/bulk/:bulkUuid/status" method="GET">
+        <service class="Magento\AsynchronousOperations\Api\BulkStatusInterface" method="getBulkShortStatus"/>
+        <resources>
+            <resource ref="Magento_Logging::system_magento_logging_bulk_operations" />
+        </resources>
+    </route>
+
+    <route url="/V1/bulk/:bulkUuid/operation-status/:status" method="GET">
+        <service class="Magento\AsynchronousOperations\Api\BulkStatusInterface" method="getOperationsCountByBulkIdAndStatus"/>
         <resources>
             <resource ref="Magento_Logging::system_magento_logging_bulk_operations" />
         </resources>
